Add slash command functionality with Discord bot
This commit finalizes the basic functionality for slash commands with
the app's Discord bot. With this commit users who have authenticated
with the website will be able to add tasks simply by using the relevant
commands of the Discord bot.

This commit also rolls back a feature wherein the application stored
Discord authentication tokens; it now only stores user IDs.